Updated the serializer to return Skills
Added a Skill serializer and called it through the SerializerMethodField
in the CustomUserSerializer. Currently returning an array of
objects(dictionaries) corresponding to each skill and limtied to just
the 'name' field, but probably ideal to re-jig this to return a straight
array of strings so it aligns with the POST json format.