Refining serializers for skills.
Thought this was working correctly in the one serialiser yesterday for
GET/POST, but discovered today that nooooo, tis not the case.

Couldn't quite figure out how to make the serializer method field only
run on the GET component, so ended up creating a new serializer for
'GET' requests (called 'CustomUserSerializerRead') and updated the
views.py with this.

Also updated the PUT request to update the skills.